// autodesk-project/src/app/model/model.ts
export interface User{
    id: number;
    firstName: string;
    lastName: string;
    userName: string;
    password: string;
  }


export interface LoginUser {
    userName: string;
    password: string;
  }


export interface UserProfile{
    id: number;
    firstName: string;
    lastName: string;
  }